+++ src/graphnet/data/__init__.py
@@ -3,14 +3,3 @@
 `graphnet.data` enables converting domain-specific data to industry-standard,
 intermediate file formats  and reading this data.
 """
-
-# Configuration
-from graphnet.utilities.imports import has_torch_package
-
-if has_torch_package():
-    import torch.multiprocessing
-    from .dataset import EnsembleDataset
-
-    torch.multiprocessing.set_sharing_strategy("file_system")
-
-del has_torch_package
