+++ src/graphnet/data/dataset/__init__.py
@@ -1,0 +1,14 @@
+"""Dataset classes for training in GraphNeT."""
+# Configuration
+from graphnet.utilities.imports import has_torch_package
+
+if has_torch_package():
+    import torch.multiprocessing
+    from .dataset import EnsembleDataset, Dataset, ColumnMissingException
+    from .parquet.parquet_dataset import ParquetDataset
+    from .sqlite.sqlite_dataset import SQLiteDataset
+    from .sqlite.sqlite_dataset_perturbed import SQLiteDatasetPerturbed
+
+    torch.multiprocessing.set_sharing_strategy("file_system")
+
+del has_torch_package
