+++ src/graphnet/data/dataset/sqlite/sqlite_dataset.py
@@ -4,7 +4,7 @@ from typing import Any, List, Optional, Tuple, Union
 import pandas as pd
 import sqlite3
 
-from graphnet.data.dataset import Dataset, ColumnMissingException
+from graphnet.data.dataset.dataset import Dataset, ColumnMissingException
 
 
 class SQLiteDataset(Dataset):
